pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "vi-rag-bench"
version = "0.1.0"
description = "Vietnamese RAG Benchmark: evaluating Native/Advanced RAG pipelines across embedding models and LLMs on Vietnamese QA datasets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index-core>=0.12",
    "llama-index-vector-stores-chroma>=0.4",
    "llama-index-embeddings-huggingface>=0.5",
    "chromadb>=1.0",
    "rouge-score>=0.1",
    "requests>=2.31",
    "pandas>=2.0",
    "numpy>=2.0",
    "tqdm>=4.67",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock>=3.14"]

[project.scripts]
vi-rag-bench = "rag_bench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
